+++ FridgeApp.Infrastructure/EF/Models/ProductReadModel.cs
@@ -7,7 +7,7 @@ namespace FridgeApp.Infrastructure.EF.Models
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public int DefaultQuantity { get; set; }
+        public ushort DefaultQuantity { get; set; }
         public int Version { get; set; }
         public DateTimeOffset CreatedDateTime { get; set; }
         public DateTimeOffset? UpdatedDateTime { get; set; }
